third_party/WebKit/LayoutTests/fast/frames/frame-focus-no-focusout-event.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<iframe></iframe>
<script>
  var iframe = document.getElementsByTagName('iframe')[0];
  var focusoutEventCount = 0;
  iframe.contentDocument.documentElement.addEventListener('focusout', function () {
     ++focusoutEventCount;
  });

  test(function() {
    iframe.contentDocument.documentElement.contentEditable = true;
    iframe.contentDocument.documentElement.focus();
    assert_equals(focusoutEventCount, 0);
  }, "Test that calling focus() on an iframe does not cause a focusout event to be fired");
</script>
</body>
</html>
